+++ go/packages/slslambda/event.go
@@ -2,14 +2,18 @@ package slslambda
 
 import (
 	"fmt"
-	"github.com/aws/aws-sdk-go/aws"
-	"go.buf.build/protocolbuffers/go/serverless/sdk-schema/serverless/instrumentation/tags/v1"
-	"go.buf.build/protocolbuffers/go/serverless/sdk-schema/serverless/instrumentation/v1"
 	"reflect"
 	"time"
+
+	"github.com/aws/aws-sdk-go/aws"
+	tagsv1 "go.buf.build/protocolbuffers/go/serverless/sdk-schema/serverless/instrumentation/tags/v1"
+	instrumentationv1 "go.buf.build/protocolbuffers/go/serverless/sdk-schema/serverless/instrumentation/v1"
 )
 
-const telemetryErrorGeneratedV1 = "telemetry.error.generated.v1"
+const (
+	telemetryErrorGeneratedV1   = "telemetry.error.generated.v1"
+	telemetryWarningGeneratedV1 = "telemetry.warning.generated.v1"
+)
 
 type (
 	errorEvent struct {
@@ -57,9 +61,9 @@ func convertToProtoErrorEvent(event errorEvent, traceID, spanID []byte, errType 
 		EventName:         telemetryErrorGeneratedV1,
 		Tags: &tagsv1.Tags{
 			Error: &tagsv1.ErrorTags{
-				Name:       errorType(event.error),
-				Message:    aws.String(event.Error()),
-				Type:       errType,
+				Name:    errorType(event.error),
+				Message: aws.String(event.Error()),
+				Type:    errType,
 			},
 		},
 	}
@@ -77,7 +81,7 @@ func convertToProtoWarningEvent(event warningEvent, traceID, spanID []byte) (*in
 		TraceId:           traceID,
 		SpanId:            spanID,
 		TimestampUnixNano: uint64(event.timestamp.UnixNano()),
-		EventName:         telemetryErrorGeneratedV1,
+		EventName:         telemetryWarningGeneratedV1,
 		Tags: &tagsv1.Tags{
 			Warning: &tagsv1.WarningTags{
 				Message: event.message,