+++ node/test/python/aws-lambda-sdk/benchmark/lib/run.js
@@ -7,7 +7,10 @@ const getProcessFunction = require('../../../../../test/lib/get-process-function
 const cleanup = require('../../lib/cleanup');
 const runBenchmark = require('../../../../../test/lib/benchmark/run');
 
-const fixturesDirname = path.resolve(__dirname, '../../fixtures/lambdas');
+const fixturesDirname = path.resolve(
+  __dirname,
+  '../../../../../../python/packages/aws-lambda-sdk/tests/fixtures/lambdas'
+);
 
 module.exports = async (useCasesConfig, coreConfig) =>
   runBenchmark({
@@ -17,11 +20,11 @@ module.exports = async (useCasesConfig, coreConfig) =>
       TracePayload,
       fixturesDirname,
       baseLambdaConfiguration: {
-        Runtime: 'nodejs16.x',
+        Runtime: 'python3.9',
         Layers: [coreConfig.layerInternalArn],
         Environment: {
           Variables: {
-            AWS_LAMBDA_EXEC_WRAPPER: '/opt/sls-sdk-node/exec-wrapper.sh',
+            AWS_LAMBDA_EXEC_WRAPPER: '/opt/sls-sdk-python/exec_wrapper.py',
           },
         },
       },