+++ python/packages/aws-lambda-sdk/serverless_aws_lambda_sdk/instrument/__init__.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 import os
 import time
+import json
 from functools import wraps
 from typing import List, Optional, Any
 import logging
@@ -68,6 +69,7 @@ class Instrumenter:
     def __init__(self):
         self.current_invocation_id = 0
         serverlessSdk._captured_events = []
+        serverlessSdk._event_emitter.on("captured-event", self._captured_event_handler)
         serverlessSdk._initialize()
         self.aws_lambda = serverlessSdk.trace_spans.aws_lambda
         if not serverlessSdk.org_id:
@@ -78,6 +80,9 @@ class Instrumenter:
                 + "or pass it with the options\n"
             )
         serverlessSdk.trace_spans.aws_lambda_initialization.close()
+
+    def _captured_event_handler(self, captured_event):
+        serverlessSdk._captured_events.append(captured_event)
 
     def _report_trace(self):
         payload_dct = {
@@ -91,6 +96,7 @@ class Instrumenter:
             },
             "spans": [s.to_protobuf_dict() for s in self.aws_lambda.spans],
             "events": [e.to_protobuf_dict() for e in serverlessSdk._captured_events],
+            "customTags": json.dumps(serverlessSdk._custom_tags),
         }
         payload = _get_payload(payload_dct)
         print(
@@ -137,6 +143,7 @@ class Instrumenter:
         del self.aws_lambda.trace_id
         del self.aws_lambda.end_time
         serverlessSdk._captured_events = []
+        serverlessSdk._custom_tags.clear()
         self.isRootSpanReset = True
 
     def instrument(self, user_handler: Handler) -> Handler: